Add files via upload
This code supplements my dissertation on 'Autoencoders in Discovering New Physics'. 
The primary functionality was implemented by myself, with the key functions 'pad_image' and 'normalize' provided by my professor for use in the project.
 - Autoencoder.py: Contains the implementation of the autoencoder model used in the project.
 - DataLoader1.py: Handles the data loading and pre-processing required for training the model.

The objective of this code is to experiment with autoencoders and investigate their potential applications in discovering patterns within physics data.